requirements.txt
fastapi
pymongo
python-socketio
python-jose
uvicorn
python-dotenv
python-multipart
